+++ lib/tezster_dart.dart
@@ -1,3 +1,6 @@
 library tezster_dart;
 
 export 'src/tezster_dart.dart';
+
+// KeyStore Model
+export 'models/key_store_model.dart';